requirements.txt
certifi==2019.11.28
chardet==3.0.4
easyimap==0.6.3
gunicorn==20.0.4
idna==2.9
PySocks==1.7.1
pyTelegramBotAPI==3.6.7
requests==2.23.0
ruamel.base==1.0.0
six==1.14.0
urllib3==1.25.8
